factor out CLI Python calls and state from CLIUtils
