+++ vcell-cli/src/main/java/org/vcell/cli/CLIStandalone.java
@@ -2,6 +2,8 @@ package org.vcell.cli;
 
 
 import cbit.vcell.parser.ExpressionException;
+import cbit.vcell.resource.OperatingSystemInfo;
+import cbit.vcell.resource.PropertyLoader;
 import cbit.vcell.solver.ode.ODESolverResultSet;
 import cbit.vcell.xml.ExternalDocInfo;
 import org.jlibsedml.*;
@@ -26,14 +28,17 @@ import java.util.*;
 
 
 public class CLIStandalone {
-	
+
     public static void main(String[] args) throws IOException {
         // We need to immediately Initialize the CLIHandler to parse CLIArgs
         CLIHandler.instantiateCliHandler(args);
         CLIHandler cliArgs = CLIHandler.getCLIHandler();
         File inputFile = new File(cliArgs.getInputFilePath());
 
-        Executable.setTimeoutMS(CLIUtils.EXECUTABLE_MAX_WALLCLOK_MILLIS);
+        PropertyLoader.loadProperties();
+        CLIPythonManager.getInstance().instantiatePythonProcess();
+
+        Executable.setTimeoutMS(CLIHandler.EXECUTABLE_MAX_WALLCLOK_MILLIS);
         
         if (cliArgs.inConversionMode()){
             CLIStandalone.wrappedConvertFiles(cliArgs);
@@ -43,7 +48,7 @@ public class CLIStandalone {
             CLIStandalone.singleMode(inputFile);
         }
 
-        CLIUtils.getCLIUtils().closePythonProcess(); // WARNING: Python will need reinstantiation after this is called
+        CLIPythonManager.getInstance().closePythonProcess(); // WARNING: Python will need reinstantiation after this is called
     }
 
     private static boolean isBatchExecution(String outputBaseDir) {
@@ -149,7 +154,6 @@ public class CLIStandalone {
 
     private static void singleExecOmex(File inputFile) throws Exception {
         int nModels, nSimulations, nSedml, nTasks, nOutputs, nReportsCount = 0, nPlots2DCount = 0, nPlots3DCount = 0;
-        CLIUtils utils = CLIUtils.getCLIUtils();
         CLIHandler cliHandler = CLIHandler.getCLIHandler();
         String inputFileName = inputFile.getAbsolutePath(), bioModelBaseName = FileUtils.getBaseName(inputFile.getName()), 
             outputDir = Paths.get(cliHandler.getOutputDirPath(), bioModelBaseName).toString(), outputBaseDir = cliHandler.getOutputDirPath(); // bioModelBaseName = input file without the path 
@@ -182,7 +186,7 @@ public class CLIStandalone {
         }
         
         CLIUtils.removeAndMakeDirs(new File(outputDir));
-        utils.generateStatusYaml(inputFileName, outputDir);	// generate Status YAML
+        CLIUtils.generateStatusYaml(inputFileName, outputDir);	// generate Status YAML
         
         // from here on, we need to collect errors, since some subtasks may succeed while other do not
         // we now have the log file created, so that we also have a place to put them
@@ -210,7 +214,7 @@ public class CLIStandalone {
                 SedML sedmlFromOmex = Libsedml.readDocument(completeSedmlPath).getSedMLModel();
 
                 String[] sedmlNameSplit;
-                sedmlNameSplit = sedmlLocation.split(CLIUtils.isWindowsPlatform ? "\\\\" : "/", -2);
+                sedmlNameSplit = sedmlLocation.split(OperatingSystemInfo.getInstance().isWindows() ? "\\\\" : "/", -2);
                 sedmlName = sedmlNameSplit[sedmlNameSplit.length - 1];
                 logOmexMessage += "Processing " + sedmlName + ". ";
 
@@ -240,7 +244,7 @@ public class CLIStandalone {
                 CLIUtils.drawBreakLine("-", 100);
 
                 // For appending data for SED Plot2D and Plot3d to HDF5 files following a temp convention
-                utils.genSedmlForSed2DAnd3D(inputFileName, outputDir);
+                CLIUtils.genSedmlForSed2DAnd3D(inputFileName, outputDir);
                 // SED-ML file generated by python VCell_cli_util
                 sedmlPathwith2dand3d = new File(String.valueOf(sedmlPath2d3d), "simulation_" + sedmlName);
                 Path path = Paths.get(sedmlPathwith2dand3d.getAbsolutePath());
@@ -261,8 +265,8 @@ public class CLIStandalone {
             	String prefix = "SED-ML processing for " + sedmlLocation + " failed with error: ";
             	logDocumentError = prefix + e.getMessage();
             	String type = e.getClass().getSimpleName();
-            	utils.setOutputMessage(sedmlLocation, sedmlName, outputDir, "sedml", logDocumentMessage);
-            	utils.setExceptionMessage(sedmlLocation, sedmlName, outputDir, "sedml", type, logDocumentError);
+            	CLIUtils.setOutputMessage(sedmlLocation, sedmlName, outputDir, "sedml", logDocumentMessage);
+            	CLIUtils.setExceptionMessage(sedmlLocation, sedmlName, outputDir, "sedml", type, logDocumentError);
                 writeDetailedErrorList(outputBaseDir, bioModelBaseName + ",  doc:    " + type + ": " + logDocumentError);
                 writeDetailedResultList(outputBaseDir, bioModelBaseName + "," + sedmlName + "," + logDocumentError);
            	
@@ -270,7 +274,7 @@ public class CLIStandalone {
                 e.printStackTrace(System.err);
                 somethingFailed = true;
                 oneSedmlDocumentFailed = true;
-                utils.updateSedmlDocStatusYml(sedmlLocation, Status.FAILED, outputDir);
+                CLIUtils.updateSedmlDocStatusYml(sedmlLocation, Status.FAILED, outputDir);
                 continue;
             }
             
@@ -296,7 +300,7 @@ public class CLIStandalone {
             	String str = "Starting simulate all tasks... ";
             	System.out.println(str);
             	logDocumentMessage += str;
-            	resultsHash = solverHandler.simulateAllTasks(utils, externalDocInfo, sedml, outDirForCurrentSedml, outputDir, outputBaseDir, sedmlLocation, keepTempFiles, exactMatchOnly);
+            	resultsHash = solverHandler.simulateAllTasks(externalDocInfo, sedml, outDirForCurrentSedml, outputDir, outputBaseDir, sedmlLocation, keepTempFiles, exactMatchOnly);
             	Map<String, String> sim2Hdf5Map = solverHandler.sim2Hdf5Map;	// may not need it
             } catch(Exception e) {
             	somethingFailed = true;
@@ -327,7 +331,7 @@ public class CLIStandalone {
         		}
         		logDocumentMessage += "Generating outputs... ";
 
-                reportsHash = utils.generateReportsAsCSV(sedml, resultsHash, outDirForCurrentSedml, outputDir, sedmlLocation);
+                reportsHash = CLIUtils.generateReportsAsCSV(sedml, resultsHash, outDirForCurrentSedml, outputDir, sedmlLocation);
 
         		if(reportsHash == null || reportsHash.size() == 0) {
         			oneSedmlDocumentFailed = true;
@@ -345,9 +349,9 @@ public class CLIStandalone {
         		}
 
         		logDocumentMessage += "Generating HDF5 file... ";
-        		String idNamePlotsMap = utils.generateIdNamePlotsMap(sedml, outDirForCurrentSedml);
+        		String idNamePlotsMap = CLIUtils.generateIdNamePlotsMap(sedml, outDirForCurrentSedml);
         		
-                utils.execPlotOutputSedDoc(inputFileName, idNamePlotsMap, outputDir);							// create the HDF5 file
+                CLIUtils.execPlotOutputSedDoc(inputFileName, idNamePlotsMap, outputDir);							// create the HDF5 file
 
         		if(!containsExtension(outputDir, "h5")) {
         			oneSedmlDocumentFailed = true;
@@ -357,17 +361,17 @@ public class CLIStandalone {
         			logDocumentMessage += "Done. ";
         		}
 
-                utils.genPlotsPseudoSedml(sedmlLocation, outDirForCurrentSedml.toString());	// generate the plots
+                CLIUtils.genPlotsPseudoSedml(sedmlLocation, outDirForCurrentSedml.toString());	// generate the plots
     			oneSedmlDocumentSucceeded = true;
         	} catch (Exception e) {
                 somethingFailed = true;
                 oneSedmlDocumentFailed = true;
             	logDocumentError += e.getMessage();
             	String type = e.getClass().getSimpleName();
-            	utils.setOutputMessage(sedmlLocation, sedmlName, outputDir, "sedml", logDocumentMessage);
-            	utils.setExceptionMessage(sedmlLocation, sedmlName, outputDir, "sedml", type, logDocumentError);
+            	CLIUtils.setOutputMessage(sedmlLocation, sedmlName, outputDir, "sedml", logDocumentMessage);
+                CLIUtils.setExceptionMessage(sedmlLocation, sedmlName, outputDir, "sedml", type, logDocumentError);
                 writeDetailedErrorList(outputBaseDir, bioModelBaseName + ",  doc:    " + type + ": " + logDocumentError);
-                utils.updateSedmlDocStatusYml(sedmlLocation, Status.FAILED, outputDir);
+                CLIUtils.updateSedmlDocStatusYml(sedmlLocation, Status.FAILED, outputDir);
                 org.apache.commons.io.FileUtils.deleteDirectory(new File(String.valueOf(sedmlPath2d3d)));	// removing temp path generated from python
                 continue;
         	}
@@ -376,10 +380,10 @@ public class CLIStandalone {
             	Exception e = new RuntimeException("Failure executing the sed document. ");
             	logDocumentError += e.getMessage();
             	String type = e.getClass().getSimpleName();
-            	utils.setOutputMessage(sedmlLocation, sedmlName, outputDir, "sedml", logDocumentMessage);
-            	utils.setExceptionMessage(sedmlLocation, sedmlName, outputDir, "sedml", type, logDocumentError);
+                CLIUtils.setOutputMessage(sedmlLocation, sedmlName, outputDir, "sedml", logDocumentMessage);
+                CLIUtils.setExceptionMessage(sedmlLocation, sedmlName, outputDir, "sedml", type, logDocumentError);
                 writeDetailedErrorList(outputBaseDir, bioModelBaseName + ",  doc:    " + type + ": " + logDocumentError);
-                utils.updateSedmlDocStatusYml(sedmlLocation, Status.FAILED, outputDir);
+                CLIUtils.updateSedmlDocStatusYml(sedmlLocation, Status.FAILED, outputDir);
                 org.apache.commons.io.FileUtils.deleteDirectory(new File(String.valueOf(sedmlPath2d3d)));	// removing temp path generated from python
                 continue;
             }
@@ -387,8 +391,8 @@ public class CLIStandalone {
 
             // archiving res files
             CLIUtils.zipResFiles(new File(outputDir));
-            utils.setOutputMessage(sedmlLocation, sedmlName, outputDir, "sedml", logDocumentMessage);
-            utils.updateSedmlDocStatusYml(sedmlLocation, Status.SUCCEEDED, outputDir);
+            CLIUtils.setOutputMessage(sedmlLocation, sedmlName, outputDir, "sedml", logDocumentMessage);
+            CLIUtils.updateSedmlDocStatusYml(sedmlLocation, Status.SUCCEEDED, outputDir);
         }
 
         omexHandler.deleteExtractedOmex();
@@ -406,16 +410,16 @@ public class CLIStandalone {
         	if(oneSedmlDocumentSucceeded) {		// some succeeded, some failed
         		error = " At least one document in this archive failed to execute";
         	}
-        	utils.updateOmexStatusYml(CLIUtils.Status.FAILED, outputDir, duration + "");
+            CLIUtils.updateOmexStatusYml(CLIUtils.Status.FAILED, outputDir, duration + "");
         	System.err.println(error);
         	logOmexMessage += error;
         	writeErrorList(outputBaseDir, bioModelBaseName);
         } else {
-        	utils.updateOmexStatusYml(CLIUtils.Status.SUCCEEDED, outputDir, duration + "");
+            CLIUtils.updateOmexStatusYml(CLIUtils.Status.SUCCEEDED, outputDir, duration + "");
         	writeFullSuccessList(outputBaseDir, bioModelBaseName);
         	logOmexMessage += " Done";
         }
-        utils.setOutputMessage("null", "null", outputDir, "omex", logOmexMessage);
+        CLIUtils.setOutputMessage("null", "null", outputDir, "omex", logOmexMessage);
         
     }
 
@@ -451,7 +455,7 @@ public class CLIStandalone {
             for (String simName : resultsHash.keySet()) {
                 String CSVFilePath = Paths.get(outDirForCurrentVcml.toString(), simName + ".csv").toString();
                 CLIUtils.createCSVFromODEResultSet(resultsHash.get(simName), new File(CSVFilePath));
-                CLIUtils.getCLIUtils().transposeVcmlCsv(CSVFilePath);
+                CLIUtils.transposeVcmlCsv(CSVFilePath);
             }
         } catch(IOException e){
             System.err.println("IOException while processing VCML " + vcmlFile.getName());
